document_processor/requirements.txt
pgvector
psycopg2-binary
langchain
openai>=1.0.0
PyPDF2
pypdf
flask
python-dotenv
langchain-community
tiktoken>=0.3.0
langchain-openai>=0.1.0
langchain-postgres>=0.1.0